Added women's rankings as a game mode!
I tried adding a combined gender rankings game mode, but it'd be really annoying to check every possible configuration of a country using their men's rankings, or their women's rankings. Maybe I'll try something like this in the future, but for now, there are only options for men's and women's rankings.